+++ tower-http/src/trace/on_response.rs
@@ -102,7 +102,7 @@ impl DefaultOnResponse {
 }
 
 impl<B> OnResponse<B> for DefaultOnResponse {
-    fn on_response(self, response: &Response<B>, latency: Duration, _: &Span) {
+    fn on_response(self, response: &Response<B>, latency: Duration, span: &Span) {
         let latency = Latency {
             unit: self.latency_unit,
             duration: latency,
@@ -112,6 +112,7 @@ impl<B> OnResponse<B> for DefaultOnResponse {
             .then(|| tracing::field::debug(response.headers()));
 
         event_dynamic_lvl!(
+            parent: span,
             self.level,
             %latency,
             status = status(response),